CTTR {koRpus}	R Documentation

Lexical diversity: Carroll's corrected TTR (CTTR)

Description

This is just a convenient wrapper function for lex.div.

Usage

CTTR(txt, char = FALSE, ...)

Arguments

txt
An object of class kRp.text containing the tagged text to be analyzed.

char
Logical, defining whether data for plotting characteristic curves should be calculated.

...
Further valid options for the main function, see lex.div for details.

Details

Calculates Carroll's corrected TTR (CTTR). In contrast to lex.div, which by default calculates all possible measures and their progressing characteristics, this function will only calculate the CTTR value, and characteristics are off by default.

Value

An object of class kRp.TTR.

See Also

kRp.POS.tags, kRp.text, kRp.TTR

Examples

## Not run: 
CTTR(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]