FOG {koRpus}	R Documentation

Readability: Gunning FOG Index

Description

This is just a convenient wrapper function for readability.

Usage

FOG(
  txt.file,
  hyphen = NULL,
  parameters = list(syll = 3, const = 0.4, suffix = c("es", "ed", "ing")),
  ...
)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

hyphen
An object of class kRp.hyphen. If NULL, the text will be hyphenated automatically.

parameters
A list with named magic numbers and a vector with verb suffixes, defining the relevant parameters for the index, or one of "PSK" or "NRI".

...
Further valid options for the main function, see readability for details.

Details

Calculates the Gunning FOG index. In contrast to readability, which by default calculates all possible indices, this function will only calculate the index value.

If parameters="PSK", the revised parameters by Powers-Sumner-Kearl (1958) are used, and if parameters="NRI", the simplified parameters from the Navy Readability Indexes, respectively.

Value

An object of class kRp.readability.

References

DuBay, W.H. (2004). The Principles of Readability. Costa Mesa: Impact Information. WWW: http://www.impact-information.com/impactinfo/readability02.pdf; 22.03.2011.

Powers, R.D, Sumner, W.A, & Kearl, B.E. (1958). A recalculation of four adult readability formulas, Journal of Educational Psychology, 49(2), 99–105.

Examples

## Not run: 
FOG(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]