HDD {koRpus}	R Documentation

Lexical diversity: HD-D (vocd-d)

Description

This is just a convenient wrapper function for lex.div.

Usage

HDD(txt, rand.sample = 42, char = FALSE, ...)

Arguments

txt
An object of class kRp.text containing the tagged text to be analyzed.

rand.sample
An integer value, how many tokens should be assumed to be drawn for calculating HD-D.

char
Logical, defining whether data for plotting characteristic curves should be calculated.

...
Further valid options for the main function, see lex.div for details.

Details

This function calculates HD-D, an idealized version of vocd-d (see McCarthy & Jarvis, 2007). In contrast to lex.div, which by default calculates all possible measures and their progressing characteristics, this function will only calculate the HD-D value, and characteristics are off by default.

Value

An object of class kRp.TTR.

References

McCarthy, P.M. & Jarvis, S. (2007). vocd: A theoretical and empirical evaluation. Language Testing, 24(4), 459–488.

See Also

kRp.POS.tags, kRp.text, kRp.TTR

Examples

## Not run: 
HDD(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]