RIX {koRpus}	R Documentation

Readability: Anderson's Readability Index (RIX)

Description

This is just a convenient wrapper function for readability.

Usage

RIX(txt.file, parameters = c(char = 6), ...)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

parameters
A numeric vector with named magic numbers, defining the relevant parameters for the index.

...
Further valid options for the main function, see readability for details.

Details

This function calculates the Readability Index (RIX) by Anderson, which is a simplified version of the l\"asbarhetsindex (LIX) by Bj\"ornsson. In contrast to readability, which by default calculates all possible indices, this function will only calculate the index value.

This formula doesn't need syllable count.

Value

An object of class kRp.readability.

References

Anderson, J. (1981). Analysing the readability of english and non-english texts in the classroom with Lix. In Annual Meeting of the Australian Reading Association, Darwin, Australia.

Anderson, J. (1983). Lix and Rix: Variations on a little-known readability index. Journal of Reading, 26(6), 490–496.

Examples

## Not run: 
  RIX(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]