SMOG {koRpus}	R Documentation

Readability: Simple Measure of Gobbledygook (SMOG)

Description

This is just a convenient wrapper function for readability.

Usage

SMOG(
  txt.file,
  hyphen = NULL,
  parameters = c(syll = 3, sqrt = 1.043, fact = 30, const = 3.1291, sqrt.const = 0),
  ...
)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

hyphen
An object of class kRp.hyphen. If NULL, the text will be hyphenated automatically.

parameters
A numeric vector with named magic numbers, defining the relevant parameters for the index.

...
Further valid options for the main function, see readability for details.

Details

This function calculates the Simple Measure of Gobbledygook (SMOG). In contrast to readability, which by default calculates all possible indices, this function will only calculate the index value.

By default calculates formula D by McLaughlin (1969). If parameters is set to SMOG="C", formula C will be calculated. If parameters is set to SMOG="simple", the simplified formula is used, and if parameters="de", the formula adapted to German texts ("Qu", Bamberger & Vanecek, 1984, p. 78).

Value

An object of class kRp.readability.

References

Bamberger, R. & Vanecek, E. (1984). Lesen–Verstehen–Lernen–Schreiben. Wien: Jugend und Volk.

McLaughlin, G.H. (1969). SMOG grading – A new readability formula. Journal of Reading, 12(8), 639–646.

Examples

## Not run: 
SMOG(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]