TRI {koRpus}	R Documentation

Readability: Kuntzsch's Text-Redundanz-Index

Description

This is just a convenient wrapper function for readability.

Usage

TRI(
  txt.file,
  hyphen = NULL,
  parameters = c(syll = 1, word = 0.449, pnct = 2.467, frgn = 0.937, const = 14.417),
  ...
)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

hyphen
An object of class kRp.hyphen. If NULL, the text will be hyphenated automatically.

parameters
A numeric vector with named magic numbers, defining the relevant parameters for the index.

...
Further valid options for the main function, see readability for details.

Details

This function calculates Kuntzsch's Text-Redundanz-Index (text redundancy index). In contrast to readability, which by default calculates all possible indices, this function will only calculate the index value.

Value

An object of class kRp.readability.

Examples

## Not run: 
  TRI(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]