available.koRpus.lang {koRpus}	R Documentation

List available language packages

Description

Get a list of all currently available language packages for koRpus from the official l10n repository.

Usage

available.koRpus.lang(repos = "https://undocumeantit.github.io/repos/l10n/")

Arguments

repos
The URL to additional repositories to query. You should probably leave this to the default, but if you would like to use a third party repository, you're free to do so. The value is temporarily appended to the repos currently returned by getOption("repos").

Details

koRpus' language support is modular by design, meaning you can (and must) load an extension package for each language you want to work with in a given session. These language support packages are named koRpus.lang.**, where ** is replaced by a valid language identifier (like en for English or de for German). See set.lang.support for more details.

This function downloads the package list from (also) the official localization repository for koRpus and lists all currently available language packages that you could install and load. Apart from than it does not download or install anything.

You can install the packages by either calling the convenient wrapper function install.koRpus.lang, or install.packages (see examples).

Value

Returns an invisible character vector with all available language packages.

See Also

install.koRpus.lang

Examples

## Not run: 
# see all available language packages
available.koRpus.lang()

# install support for German
install.koRpus.lang("de")
# alternatively, you could call install.packages directly
install.packages("koRpus.lang.de", repos="https://undocumeantit.github.io/repos/l10n/")

## End(Not run)

[Package koRpus version 0.13-8 Index]