bormuth {koRpus}	R Documentation

Readability: Bormuth's Mean Cloze and Grade Placement

Description

This is just a convenient wrapper function for readability.

Usage

bormuth(txt.file, word.list, clz=35,
   meanc=c(const=0.886593, awl=0.08364, afw=0.161911,
     asl1=0.021401, asl2=0.000577, asl3=0.000005),
   grade=c(const=4.275, m1=12.881, m2=34.934, m3=20.388,
     c1=26.194, c2=2.046, c3=11.767, mc1=44.285, mc2=97.62,
     mc3=59.538), ...)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

word.list
A vector or matrix (with exactly one column) which defines familiar words. For valid results the long Dale-Chall list with 3000 words should be used.

clz
Integer, the cloze criterion score in percent.

meanc
A numeric vector with named magic numbers, defining the relevant parameters for Mean Cloze calculation.

grade
A numeric vector with named magic numbers, defining the relevant parameters for Grade Placement calculation. If omitted, Grade Placement will not be calculated.

...
Further valid options for the main function, see readability for details.

Details

Calculates Bormuth's Mean Cloze and estimted grade placement. In contrast to readability, which by default calculates all possible indices, this function will only calculate the index value.

This formula doesn't need syllable count.

Value

An object of class kRp.readability.

Examples

## Not run: 
  bormuth(tagged.text, word.list=new.dale.chall.wl)

## End(Not run)

[Package koRpus version 0.13-8 Index]