danielson.bryan {koRpus}	R Documentation

Readability: Danielson-Bryan

Description

This is just a convenient wrapper function for readability.

Usage

danielson.bryan(
  txt.file,
  db1 = c(cpb = 1.0364, cps = 0.0194, const = 0.6059),
  db2 = c(const = 131.059, cpb = 10.364, cps = 0.194),
  ...
)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

db1
A numeric vector with named magic numbers, defining the relevant parameters for the first formula (regression).

db2
A numeric vector with named magic numbers, defining the relevant parameters for the second formula (cloze equivalent).

...
Further valid options for the main function, see readability for details.

Details

Calculates the two Danielson-Bryan formulas. In contrast to readability, which by default calculates all possible indices, this function will only calculate the index value.

This formula doesn't need syllable count.

Value

An object of class kRp.readability.

Examples

## Not run: 
  danielson.bryan(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]