flesch {koRpus}	R Documentation

Readability: Flesch Readability Ease

Description

This is just a convenient wrapper function for readability.

Usage

flesch(
  txt.file,
  hyphen = NULL,
  parameters = c(const = 206.835, asl = 1.015, asw = 84.6),
  ...
)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

hyphen
An object of class kRp.hyphen. If NULL, the text will be hyphenated automatically.

parameters
Either a numeric vector with named magic numbers, defining the relevant parameters for the index, or a valid character string naming a preset for implemented languages ("de", "es", "es-s", "nl", "nl-b", "fr").

...
Further valid options for the main function, see readability for details.

Details

Calculates the Flesch Readability Ease index. In contrast to readability, which by default calculates all possible indices, this function will only calculate the Flesch RE value.

Certain internationalisations of the parameters are also implemented. They can be used by setting parameters to "es" (Fernandez-Huerta), "es-s" (Szigriszt), "nl" (Douma), "nl-b" (Brouwer), "de" (Amstad) or "fr" (Kandel-Moles). If parameters="PSK", the revised parameters by Powers-Sumner-Kearl (1958) are used to calculate a grade level.

Value

An object of class kRp.readability.

See Also

flesch.kincaid for grade levels, farr.jenkins.paterson for a simplified Flesch formula.

Examples

## Not run: 
flesch(german.tagged.text, parameters="de")

## End(Not run)

[Package koRpus version 0.13-8 Index]