fucks {koRpus}	R Documentation

Readability: Fucks' Stilcharakteristik

Description

This is just a convenient wrapper function for readability.

Usage

fucks(txt.file, ...)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

...
Further valid options for the main function, see readability for details.

Details

Calculates Fucks' Stilcharakteristik ("characteristics of style"; Fucks, 1955, as cited in Briest, 1974). In contrast to readability, which by default calculates all possible indices, this function will only calculate the index value.

Value

An object of class kRp.readability.

References

Briest, W. (1974). Kann man Verständlichkeit messen? Zeitschrift für Phonetik, Sprachwissenschaft und Kommunikationsforschung, 27, 543–563.

Examples

## Not run: 
  fucks(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]