get.kRp.env {koRpus}	R Documentation

Get koRpus session settings

Description

The function get.kRp.env returns information on your session environment regarding the koRpus package, e.g. where your local TreeTagger installation resides, if it was set before using set.kRp.env.

Usage

get.kRp.env(..., errorIfUnset = TRUE)

Arguments

...
Named parameters to get from the koRpus environment. Valid arguments are:

TT.cmd

Logical, whether the set tagger command should be returned.

lang

Logical, whether the set language should be returned.

TT.options

Logical, whether the set TT.options for treetag should be returned.

hyph.cache.file

Logical, whether the set hyphenation cache file for hyphen should be returned.

add.desc

Logical, whether tag descriptions should be added directly to tagged text objects.

errorIfUnset
Logical, if TRUE and the desired property is not set at all, the function will fail with an error message.

Details

For the most part, get.kRp.env is a convenient wrapper for getOption.

Value

A character string or list, possibly including:

TT.cmd
Path information for the TreeTagger command

lang
The specified language

TT.options
A list with options for treetag

hyph.cache.file
The specified hyphenation cache file for hyphen

See Also

set.kRp.env

Examples

set.kRp.env(lang="en")
get.kRp.env(lang=TRUE)

[Package koRpus version 0.13-8 Index]