kRp.lang,-class {koRpus}	R Documentation

S4 Class kRp.lang

Description

This class is used for objects that are returned by guess.lang.

Slots

lang

A character string, naming the language (by its ISO 639-3 identifier) that was estimated for the analized text in this object.

lang.name

A character string, full name of the estimated language.

txt

A character string containing the analized part of the text.

txt.full

A character string containing the full text.

udhr

A data.frame with full analysis results for each language tried.

Contructor function

Should you need to manually generate objects of this class (which should rarely be the case), the contructor function kRp_lang(...) can be used instead of new("kRp.lang", ...).

[Package koRpus version 0.13-8 Index]