nWS {koRpus}	R Documentation

Readability: Neue Wiener Sachtextformeln

Description

This is just a convenient wrapper function for readability.

Usage

nWS(
  txt.file,
  hyphen = NULL,
  parameters = c(ms.syll = 3, iw.char = 6, es.syll = 1),
  nws1 = c(ms = 19.35, sl = 0.1672, iw = 12.97, es = 3.27, const = 0.875),
  nws2 = c(ms = 20.07, sl = 0.1682, iw = 13.73, const = 2.779),
  nws3 = c(ms = 29.63, sl = 0.1905, const = 1.1144),
  nws4 = c(ms = 27.44, sl = 0.2656, const = 1.693),
  ...
)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

hyphen
An object of class kRp.hyphen. If NULL, the text will be hyphenated automatically.

parameters
A numeric vector with named magic numbers, defining the relevant parameters for all formulas of the index.

nws1
A numeric vector with named magic numbers for the first of the formulas.

nws2
A numeric vector with named magic numbers for the second of the formulas.

nws3
A numeric vector with named magic numbers for the third of the formulas.

nws4
A numeric vector with named magic numbers for the fourth of the formulas.

...
Further valid options for the main function, see readability for details.

Details

This function calculates the new Wiener Sachtextformeln (formulas 1 to 4). In contrast to readability, which by default calculates all possible indices, this function will only calculate the index values.

Value

An object of class kRp.readability.

References

Bamberger, R. & Vanecek, E. (1984). Lesen–Verstehen–Lernen–Schreiben. Wien: Jugend und Volk.

Examples

## Not run: 
nWS(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]