plot {koRpus}	R Documentation

Plot method for objects of class kRp.text

Description

Plot method for S4 objects of class kRp.text, plots the frequencies of tagged word classes.

Usage

plot(x, y, ...)

## S4 method for signature 'kRp.text,missing'
plot(x, what = "wclass", ...)

Arguments

x
An object of class kRp.text

y
From the generic plot function, ignored for koRpus class objects.

...
Any other argument suitable for plot()

what
Character string, valid options are:

"wclass":

Barplot of distribution of word classes

"letters":

Line plot of distribution of word length in letters

See Also

kRp.text

Examples

# code is only run when the english language package can be loaded
if(require("koRpus.lang.en", quietly = TRUE)){
  sample_file <- file.path(
    path.package("koRpus"), "examples", "corpus", "Reality_Winner.txt"
  )
  tokenized.obj <- tokenize(
    txt=sample_file,
    lang="en"
  )
  plot(tokenized.obj)
} else {}

[Package koRpus version 0.13-8 Index]