traenkle.bailer {koRpus}	R Documentation

Readability: Traenkle-Bailer Formeln

Description

This is just a convenient wrapper function for readability.

Usage

traenkle.bailer(
  txt.file,
  TB1 = c(const = 224.6814, awl = 79.8304, asl = 12.24032, prep = 1.292857),
  TB2 = c(const = 234.1063, awl = 96.11069, prep = 2.05444, conj = 1.02805),
  ...
)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

TB1
A numeric vector with named magic numbers for the first of the formulas.

TB2
A numeric vector with named magic numbers for the second of the formulas.

...
Further valid options for the main function, see readability for details.

Details

This function calculates the two formulae by Tr\"ankle-Bailer, which are based on the Dickes-Steiwer formulae. In contrast to readability, which by default calculates all possible indices, this function will only calculate the index values.

This formula doesn't need syllable count.

Value

An object of class kRp.readability.

Examples

## Not run: 
  traenkle.bailer(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]