wheeler.smith {koRpus}	R Documentation

Readability: Wheeler-Smith Score

Description

This is just a convenient wrapper function for readability.

Usage

wheeler.smith(txt.file, hyphen = NULL, parameters = c(syll = 2), ...)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

hyphen
An object of class kRp.hyphen. If NULL, the text will be hyphenated automatically.

parameters
A numeric vector with named magic numbers, defining the relevant parameters for the index.

...
Further valid options for the main function, see readability for details.

Details

This function calculates the Wheeler-Smith Score. In contrast to readability, which by default calculates all possible indices, this function will only calculate the index value.

If parameters="de", the calculation stays the same, but grade placement is done according to Bamberger & Vanecek (1984), that is for german texts.

Value

An object of class kRp.readability.

References

Bamberger, R. & Vanecek, E. (1984). Lesen–Verstehen–Lernen–Schreiben. Wien: Jugend und Volk.

Wheeler, L.R. & Smith, E.H. (1954). A practical readability formula for the classroom teacher in the primary grades. Elementary English, 31, 397–399.

Examples

## Not run: 
  wheeler.smith(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]