ARI {koRpus}	R Documentation

Readability: Automated Readability Index (ARI)

Description

This is just a convenient wrapper function for readability.

Usage

ARI(txt.file, parameters = c(asl = 0.5, awl = 4.71, const = 21.43), ...)

Arguments

txt.file
Either an object of class kRp.text, a character vector which must be be a valid path to a file containing the text to be analyzed, or a list of text features. If the latter, calculation is done by readability.num.

parameters
A numeric vector with named magic numbers, defining the relevant parameters for the index.

...
Further valid options for the main function, see readability for details.

Details

Calculates the Automated Readability Index (ARI). In contrast to readability, which by default calculates all possible indices, this function will only calculate the index value.

If parameters="NRI", the simplified parameters from the Navy Readability Indexes are used, if set to ARI="simple", the simplified formula is calculated.

This formula doesn't need syllable count.

Value

An object of class kRp.readability.

References

DuBay, W.H. (2004). The Principles of Readability. Costa Mesa: Impact Information. WWW: http://www.impact-information.com/impactinfo/readability02.pdf; 22.03.2011.

Smith, E.A. & Senter, R.J. (1967). Automated readability index. AMRL-TR-66-22. Wright-Paterson AFB, Ohio: Aerospace Medical Division.

Examples

## Not run: 
ARI(tagged.text)

## End(Not run)

[Package koRpus version 0.13-8 Index]